Refactor D-FINE module structure by removing obsolete detector file and reorganizing imports
