+++ src/otx/algo/detection/backbones/hgnetv2.py
