--- src/otx/algo/detection/detectors/dfine/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (C) 2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-
-"""Modified from D-FINE (https://github.com/Peterande/D-FINE)."""
