+++ src/otx/algo/detection/losses/__init__.py
@@ -4,6 +4,7 @@
 """Custom OTX Losses for Object Detection."""
 
 from .atss_loss import ATSSCriterion
+from .dfine_loss import DFINECriterion
 from .rtdetr_loss import DetrCriterion
 from .rtmdet_loss import RTMDetCriterion
 from .ssd_loss import SSDCriterion
@@ -17,4 +18,5 @@ __all__ = [
     "SSDCriterion",
     "YOLOv9Criterion",
     "YOLOXCriterion",
+    "DFINECriterion",
 ]
